requirements.txt
Flask==2.3.3
graphene==3.2
neo4j==5.10.0
python-dotenv==1.0.0
flask-cors==4.0.0
